pyproject.toml
[project]
name = "gmail-followup-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.170.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "plotly>=6.1.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
